+++ js/data.js
@@ -1,0 +1,49 @@
+const DESCRIPTIONS = [
+  'Вспоминая лето',
+  'Бой быков',
+  'Мы с друзьями',
+  'Утро в саду',
+  'Вот так бывает, когда забыл закрыть балкон',
+  'Просто фото',
+  'На работе бывает вот так',
+  'Моя любовь',
+  'Это памятник вождю всех времен и народов',
+  'Тачка на прокачку',
+  'Байкал',
+  'Виды города',
+  'Бульон на говядине',
+  'Снежное утро',
+  'Мой любимый город',
+  'Когда я не понял, как так',
+  'Цирк',
+  'Вспоминая прошлое',
+  'Зашел домой блин',
+  'Кто это я не помню',
+  'Реальность',
+  'Звездное небо',
+  'Моя бывшая школа',
+  'Тратуар',
+  'Дерево'
+];
+
+const NAMES = [
+  'Артём',
+  'Иван',
+  'Екатерина',
+  'Сергей',
+  'Игорь',
+  'Владимир',
+];
+
+const COMMENTARIES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+export {DESCRIPTIONS};
+export {NAMES};
+export {COMMENTARIES};
