+++ package.json
@@ -1,5 +1,6 @@
 {
   "name": "kekstagram",
+  "type": "module",
   "version": "28.0.0",
   "private": true,
   "description": "Личный проект «Кекстаграм» от HTML Academy",
